(Port to C++) command palette, dialog about; 7z; icons; fixing layouts  (#750)
* dialog about

* command palette

* fixed my test not compiling

* fix p7za_archive didn't work with 'space' in passwords and paths; remove 7z standard output but leave error output

* remove search options from global space

* fixed warning

* convert shortcuts to label to show them

* fixed rendering icons in dialog tree views

* fixing node dialog layout

* get more proper icons for code/syntax

* fix layout in table dlg

* fixing layout in pref dlg

* fixed message dialog titles

* fixed str::replace

+++ future/src/ct/ct_actions_others.cc
@@ -588,7 +588,7 @@ void CtActions::table_row_copy()
 {
     auto table_state = std::dynamic_pointer_cast<CtAnchoredWidgetState_Table>(curr_table_anchor->get_state());
     // remove rows after current
-    while (table_state->rows.size() > curr_table_anchor->current_row() + 1)
+    while ((int)table_state->rows.size() > curr_table_anchor->current_row() + 1)
         table_state->rows.pop_back();
     // remove rows between current and header
     while (table_state->rows.size() > 2)
